Additonal file renaming & function rearrangement
