+++ YARG.Core/Song/Entries/Ini/SongEntry.IniBase.cs
@@ -186,106 +186,4 @@ namespace YARG.Core.Song
             return fivelane ? DrumsType.FiveLane : DrumsType.FourLane;
         }
     }
-
-    public partial struct SongMetadata
-    {
-        public SongMetadata(AvailableParts parts, HashWrapper hash, IniSection section, string defaultPlaylist)
-        {
-            Parts = parts;
-            Hash = hash;
-            ParseSettings = ParseSettings.Default;
-            ParseSettings.DrumsType = parts.GetDrumType();
-
-            section.TryGet("name", out Name, DEFAULT_NAME);
-            section.TryGet("artist", out Artist, DEFAULT_ARTIST);
-            section.TryGet("album", out Album, DEFAULT_ALBUM);
-            section.TryGet("genre", out Genre, DEFAULT_GENRE);
-
-            if (!section.TryGet("charter", out Charter, DEFAULT_CHARTER))
-            {
-                section.TryGet("frets", out Charter, DEFAULT_CHARTER);
-            }
-
-            section.TryGet("icon", out Source, DEFAULT_SOURCE);
-            section.TryGet("playlist", out Playlist, defaultPlaylist);
-
-            _unmodifiedYear = DEFAULT_YEAR;
-            _parsedYear = DEFAULT_YEAR;
-            _intYear = int.MaxValue;
-
-            _songLength = 0;
-            _songOffset = 0;
-
-            _previewStart = 0;
-            _previewEnd = 0;
-
-            _videoStartTime = 0;
-            _videoEndTime = -1;
-
-            section.TryGet("loading_phrase", out LoadingPhrase);
-
-            if (!section.TryGet("playlist_track", out PlaylistTrack))
-                PlaylistTrack = -1;
-
-            if (!section.TryGet("album_track", out AlbumTrack))
-                AlbumTrack = -1;
-
-            section.TryGet("song_length", out _songLength);
-
-            section.TryGet("video_start_time", out _videoStartTime);
-            _videoEndTime = section.TryGet("video_end_time", out long videoEndTime) ? videoEndTime : -1000;
-
-            if (!section.TryGet("hopo_frequency", out ParseSettings.HopoThreshold))
-                ParseSettings.HopoThreshold = -1;
-
-            if (!section.TryGet("hopofreq", out ParseSettings.HopoFreq_FoF))
-                ParseSettings.HopoFreq_FoF = -1;
-
-            section.TryGet("eighthnote_hopo", out ParseSettings.EighthNoteHopo);
-
-            if (!section.TryGet("sustain_cutoff_threshold", out ParseSettings.SustainCutoffThreshold))
-                ParseSettings.SustainCutoffThreshold = -1;
-
-            if (!section.TryGet("multiplier_note", out ParseSettings.StarPowerNote))
-                ParseSettings.StarPowerNote = -1;
-
-            IsMaster = !section.TryGet("tags", out string tag) || tag.ToLower() != "cover";
-
-            if (section.TryGet("year", out _unmodifiedYear))
-            {
-                Year = _unmodifiedYear;
-            }
-            else if (section.TryGet("year_chart", out _unmodifiedYear))
-            {
-                if (_unmodifiedYear.StartsWith(", "))
-                    Year = _unmodifiedYear[2..];
-                else if (_unmodifiedYear.StartsWith(','))
-                    Year = _unmodifiedYear[1..];
-                else
-                    Year = _unmodifiedYear;
-            }
-            else
-                _unmodifiedYear = DEFAULT_YEAR;
-
-            if (!section.TryGet("preview", out _previewStart, out _previewEnd))
-            {
-                if (!section.TryGet("preview_start_time", out _previewStart) &&
-                    section.TryGet("previewStart", out double previewStartSeconds))
-                    PreviewStartSeconds = previewStartSeconds;
-
-                if (!section.TryGet("preview_end_time", out _previewEnd) &&
-                    section.TryGet("previewEnd", out double previewEndSeconds))
-                    PreviewEndSeconds = previewEndSeconds;
-            }
-
-
-            if (!section.TryGet("delay", out _songOffset) || _songOffset == 0)
-            {
-                if (section.TryGet("offset", out double songOffsetSeconds))
-                {
-                    SongOffsetSeconds = songOffsetSeconds;
-                }
-            }
-        }
-    }
 }
