+++ YARG.Core/Song/Entries/Ini/SongEntry.Sng.cs
