+++ YARG.Core/Song/Entries/Ini/SongEntry.UnpackedIni.cs
